identity_service/requirements.txt
fastapi==0.95.2
hazelcast-python-client==5.2.0
numpy==1.24.3
pydantic==1.10.8
requests==2.31.0
uvicorn==0.22.0
kafka-python==2.0.2
